% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_ln.R
\name{census_ln}
\alias{census_ln}
\title{Impute Race Based on Census Last Name Datasets}
\usage{
census_ln(surname = NULL)
}
\arguments{
\item{surname}{Required; vector of strings}
}
\value{
Data frame with all the details that are in the census datasets
}
\description{
Includes data from 2000 and 2010 census datasets
}
\examples{
census_ln(surname = "Smith")

}
