% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicolor_data.R
\docType{data}
\name{cs2000}
\alias{cs2000}
\title{2000 Census Last Name Dataset}
\format{A data frame with 151671 observations and 11 variables}
\source{
2000 Census Last Name dataset,
  \url{http://www.census.gov/topics/population/genealogy/data/2000_surnames.html}
}
\usage{
cs2000
}
\description{
2000 Census Last Name Dataset
}
\keyword{datasets}

