% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicolor_data.R
\docType{data}
\name{cs2010}
\alias{cs2010}
\title{2010 Census Last Name Dataset}
\format{A data frame with 162254 observations and 11 variables}
\source{
2010 Census Last Name dataset,
  \url{http://www.census.gov/topics/population/genealogy/data/2010_surnames.html}
}
\usage{
cs2010
}
\description{
2010 Census Last Name Dataset
}
\keyword{datasets}
