% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_surname.R
\name{cs_surname}
\alias{cs_surname}
\title{Impute Race Based on Census Last Name Datasets}
\usage{
cs_surname(surname = NULL)
}
\arguments{
\item{surname}{Required; vector of strings}
}
\value{
Data frame with all the details that are in the census datasets
}
\description{
Includes data from 2000 and 2010 census datasets
}
\examples{
\dontrun{
cs_surname(surname="Smith")
}
}

