% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicolor.R
\docType{package}
\name{ethnicolor-package}
\alias{ethnicolor-package}
\alias{ethnicolor}
\title{ethnicolor: Impute race based on names}
\description{
Impute race and ethnicity based on last name or the full name using census data, voter registration data, and data from Wikipedia. The package exposes validated ML models to predict race and ethnicity.
}
\author{
Gaurav Sood

Pete Mohanty
}
