% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_fl_reg_ln.R
\name{pred_fl_reg_ln}
\alias{pred_fl_reg_ln}
\title{Last name classifier based on Florida Voter Registration Data}
\usage{
pred_fl_reg_ln(surname = NULL)
}
\arguments{
\item{surname}{required; string; vector of surnames}
}
\value{
data.frame with original list and predicted probability the person is 
white, asian, black, and hispanic
}
\description{
Uses a validated ML model that uses bi- and tri-char tokens in the last name
to predict probability that the person is white, black, asian, and hispanic. For
more information see \url{https://github.com/soodoku/ethnicolor/tree/master/data-raw/fl_voter_reg}
}
\examples{
pred_census_ln("Smith")

}
