% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cs_surname.R
\name{print.census_ln}
\alias{print.census_ln}
\title{Print a census_ln object}
\usage{
\method{print}{census_ln}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{...}{Further arguments to passed on to other print methods.}
}
\description{
Print a census_ln object
}
