% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_ln.R
\name{census_ln}
\alias{census_ln}
\title{Impute Race Based on Census Last Name Datasets}
\usage{
census_ln(surname = NULL)
}
\arguments{
\item{surname}{Required; vector of strings}
}
\value{
Data frame with all the details that are in the census datasets
}
\description{
Appends data on percentage whites, blacks, asians, hispanics, 
mixed-race, and native americans with the last name from the 
2000 and 2010 census datasets
}
\details{
See more at: \url{https://github.com/soodoku/ethnicolor/tree/master/data-raw/census}
}
\examples{
census_ln(surname = "Smith")
top5 <- census_ln(c("Smith", "Johnson", "Williams", "Brown", "Jones"))
sum(top5$cs2010_count)
}
