% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethnicolor.R
\docType{package}
\name{ethnicolor-package}
\alias{ethnicolor}
\alias{ethnicolor-package}
\title{ethnicolor: Impute race based on names}
\description{
Impute race based on names using data from census, voter registration files etc.
}
\author{
Gaurav Sood
}

